third_party/blink/web_tests/external/wpt/webrtc-encoded-transform/script-transform-sendKeyFrameRequest.https-expected.txt

This is a testharness.js-based test.
[FAIL] sendKeyFrameRequest resolves for video receiver, and throws for video sender
  assert_equals: expected "success" but got "failure: TypeError"
[FAIL] sendKeyFrameRequest throws for audio sender/receiver
  assert_equals: expected "failure: InvalidStateError" but got "failure: TypeError"
[FAIL] sendKeyFrameRequest throws for unused transforms
  assert_equals: expected "failure: InvalidStateError" but got "failure: TypeError"
[FAIL] sendKeyFrameRequest throws for unset transforms
  assert_equals: expected "success" but got "failure: TypeError"
[FAIL] sendKeyFrameRequest rejects when the receiver is negotiated inactive, and resumes succeeding when negotiated back to active
  assert_equals: expected "success" but got "failure: TypeError"
[FAIL] sendKeyFrameRequest rejects when the receiver is stopped
  assert_equals: expected "success" but got "failure: TypeError"
Harness: the test ran to completion.
